ai-image-classifier/requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
flask-cors>=4.0
pyinstaller>=6.4.0
